+++ src/Component/Command/Output.php
@@ -7,11 +7,12 @@
 
 namespace Ulrack\Command\Component\Command;
 
-use GrizzIt\Task\Common\TaskListInterface;
 use Ulrack\Cli\Common\Io\WriterInterface;
+use GrizzIt\Task\Common\TaskListInterface;
 use Ulrack\Cli\Common\Theme\ThemeInterface;
-use Ulrack\Command\Common\Command\OutputInterface;
 use Ulrack\Cli\Common\Factory\IoFactoryInterface;
+use Ulrack\Command\Common\Command\OutputModeEnum;
+use Ulrack\Command\Common\Command\OutputInterface;
 use Ulrack\Cli\Common\Factory\ElementFactoryInterface;
 use Ulrack\Cli\Common\Generator\FormGeneratorInterface;
 
@@ -46,6 +47,13 @@ class Output implements OutputInterface
     private $elementFactory;
 
     /**
+     * Determines what information will be displayed.
+     *
+     * @var OutputModeEnum
+     */
+    private $outputMode;
+
+    /**
      * Constructor
      *
      * @param FormGeneratorInterface $formGenerator
@@ -63,6 +71,7 @@ class Output implements OutputInterface
         $this->writer = $ioFactory->createStandardWriter();
         $this->theme = $theme;
         $this->elementFactory = $elementFactory;
+        $this->outputMode = OutputModeEnum::OUTPUT_MODE_NORMAL();
     }
 
     /**
@@ -70,15 +79,21 @@ class Output implements OutputInterface
      *
      * @param string $input
      * @param string $style
+     * @param bool $verbose
      *
      * @return void
      */
-    public function write(string $input, string $style = 'text'): void
-    {
-        $style = $this->theme->getStyle($style);
-        $style->apply();
-        $this->writer->write($input);
-        $style->reset();
+    public function write(
+        string $input,
+        string $style = 'text',
+        bool $verbose = false
+    ): void {
+        if ($this->canOutput($verbose)) {
+            $style = $this->theme->getStyle($style);
+            $style->apply();
+            $this->writer->write($input);
+            $style->reset();
+        }
     }
 
     /**
@@ -86,15 +101,21 @@ class Output implements OutputInterface
      *
      * @param string $input
      * @param string $style
+     * @param bool $verbose
      *
      * @return void
      */
-    public function writeLine(string $input, string $style = 'text'): void
-    {
-        $style = $this->theme->getStyle($style);
-        $style->apply();
-        $this->writer->writeLine($input);
-        $style->reset();
+    public function writeLine(
+        string $input,
+        string $style = 'text',
+        bool $verbose = false
+    ): void {
+        if ($this->canOutput($verbose)) {
+            $style = $this->theme->getStyle($style);
+            $style->apply();
+            $this->writer->writeLine($input);
+            $style->reset();
+        }
     }
 
     /**
@@ -103,15 +124,21 @@ class Output implements OutputInterface
      *
      * @param string $input
      * @param string $style
+     * @param bool $verbose
      *
      * @return void
      */
-    public function overWrite(string $input, string $style = 'text'): void
-    {
-        $style = $this->theme->getStyle($style);
-        $style->apply();
-        $this->writer->overWrite($input);
-        $style->reset();
+    public function overWrite(
+        string $input,
+        string $style = 'text',
+        bool $verbose = false
+    ): void {
+        if ($this->canOutput($verbose)) {
+            $style = $this->theme->getStyle($style);
+            $style->apply();
+            $this->writer->overWrite($input);
+            $style->reset();
+        }
     }
 
     /**
@@ -130,19 +157,23 @@ class Output implements OutputInterface
      * @param string $content
      * @param bool $newLine
      * @param string $styleKey
+     * @param bool $verbose
      *
      * @return void
      */
     public function outputText(
         string $content,
         bool $newLine = true,
-        string $styleKey = 'text'
+        string $styleKey = 'text',
+        bool $verbose = false
     ): void {
-        $this->elementFactory->createText(
-            $content,
-            $newLine,
-            $styleKey
-        )->render();
+        if ($this->canOutput($verbose)) {
+            $this->elementFactory->createText(
+                $content,
+                $newLine,
+                $styleKey
+            )->render();
+        }
     }
 
     /**
@@ -154,6 +185,7 @@ class Output implements OutputInterface
      * @param string $style
      * @param string $boxStyle
      * @param string $keyStyle
+     * @param bool $verbose
      *
      * @return void
      */
@@ -163,16 +195,19 @@ class Output implements OutputInterface
         string $tableCharacters = 'table-characters',
         string $style = 'table-style',
         string $boxStyle = 'table-box-style',
-        string $keyStyle = 'table-key-style'
+        string $keyStyle = 'table-key-style',
+        bool $verbose = false
     ): void {
-        $this->elementFactory->createTable(
-            $keys,
-            $items,
-            $tableCharacters,
-            $style,
-            $boxStyle,
-            $keyStyle
-        )->render();
+        if ($this->canOutput($verbose)) {
+            $this->elementFactory->createTable(
+                $keys,
+                $items,
+                $tableCharacters,
+                $style,
+                $boxStyle,
+                $keyStyle
+            )->render();
+        }
     }
 
     /**
@@ -180,17 +215,21 @@ class Output implements OutputInterface
      *
      * @param array $items
      * @param string $style
+     * @param bool $verbose
      *
      * @return void
      */
     public function outputList(
         array $items,
-        string $style = 'list'
+        string $style = 'list',
+        bool $verbose = false
     ): void {
-        $this->elementFactory->createList(
-            $items,
-            $style
-        )->render();
+        if ($this->canOutput($verbose)) {
+            $this->elementFactory->createList(
+                $items,
+                $style
+            )->render();
+        }
     }
 
     /**
@@ -199,19 +238,23 @@ class Output implements OutputInterface
      * @param array $items
      * @param string $keyStyle
      * @param string $descriptionStyle
+     * @param bool $verbose
      *
      * @return void
      */
     public function outputExplainedList(
         array $items,
         string $keyStyle = 'explained-list-key',
-        string $descriptionStyle = 'explained-list-description'
+        string $descriptionStyle = 'explained-list-description',
+        bool $verbose = false
     ): void {
-        $this->elementFactory->createExplainedList(
-            $items,
-            $keyStyle,
-            $descriptionStyle
-        )->render();
+        if ($this->canOutput($verbose)) {
+            $this->elementFactory->createExplainedList(
+                $items,
+                $keyStyle,
+                $descriptionStyle
+            )->render();
+        }
     }
 
     /**
@@ -221,6 +264,7 @@ class Output implements OutputInterface
      * @param string $style
      * @param string $padding
      * @param string $margin
+     * @param bool $verbose
      *
      * @return void
      */
@@ -228,14 +272,17 @@ class Output implements OutputInterface
         string $content,
         string $style = 'block',
         string $padding = 'block-padding',
-        string $margin = 'block-margin'
+        string $margin = 'block-margin',
+        bool $verbose = false
     ): void {
-        $this->elementFactory->createBlock(
-            $content,
-            $style,
-            $padding,
-            $margin
-        )->render();
+        if ($this->canOutput($verbose)) {
+            $this->elementFactory->createBlock(
+                $content,
+                $style,
+                $padding,
+                $margin
+            )->render();
+        }
     }
 
     /**
@@ -245,6 +292,7 @@ class Output implements OutputInterface
      * @param string $progressCharacters
      * @param string $textStyle
      * @param string $progressStyle
+     * @param bool $verbose
      *
      * @return void
      */
@@ -252,13 +300,51 @@ class Output implements OutputInterface
         TaskListInterface $taskList,
         string $progressCharacters = 'progress-characters',
         string $textStyle = 'progress-text',
-        string $progressStyle = 'progress-bar'
+        string $progressStyle = 'progress-bar',
+        bool $verbose = false
     ): void {
-        $this->elementFactory->createProgressBar(
-            $taskList,
-            $progressCharacters,
-            $textStyle,
-            $progressStyle
-        )->render();
+        if ($this->canOutput($verbose)) {
+            $this->elementFactory->createProgressBar(
+                $taskList,
+                $progressCharacters,
+                $textStyle,
+                $progressStyle
+            )->render();
+        }
+    }
+
+    /**
+     * Sets the output mode.
+     *
+     * @param OutputModeEnum $outputMode
+     *
+     * @return void
+     */
+    public function setOutputMode(OutputModeEnum $outputMode): void
+    {
+        $this->outputMode = $outputMode;
+    }
+
+    /**
+     * Determines whether the output can be rendered.
+     *
+     * @param bool $verbose
+     *
+     * @return bool
+     */
+    private function canOutput(bool $verbose): bool
+    {
+        if ((string) $this->outputMode !== OutputModeEnum::OUTPUT_MODE_QUIET) {
+            if (
+                $verbose &&
+                (string) $this->outputMode !== OutputModeEnum::OUTPUT_MODE_VERBOSE
+            ) {
+                return true;
+            } elseif (!$verbose) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }