+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 - 2020-08-16
+### Added
+- Ability to output additional context.
+- Ability to suppress all output.
+- Ability to select an output mode.
+
 ## 1.0.5 - 2020-06-08
 ### Fixed
 - Compatibility with ulrack/services version 3.0
@@ -29,7 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created the initial implementation of the command package.
 
 # Versions
-- [1.0.5 > Unreleased](https://github.com/ulrack/command/compare/1.0.5...HEAD)
+- [2.0.0 > Unreleased](https://github.com/ulrack/command/compare/2.0.0...HEAD)
+- [1.0.5 > 2.0.0](https://github.com/ulrack/command/compare/1.0.5...2.0.0)
 - [1.0.4 > 1.0.5](https://github.com/ulrack/command/compare/1.0.4...1.0.5)
 - [1.0.3 > 1.0.4](https://github.com/ulrack/command/compare/1.0.3...1.0.4)
 - [1.0.2 > 1.0.3](https://github.com/ulrack/command/compare/1.0.2...1.0.3)